Large update, read description
Adjusted the shaders for compatibility with 1.13 Water Mechanics Backport mod.

Altered the Neat config so that Artifacts' mimics won't display their health bar.

Altered the Quark config so that cave roots are enabled and the stone revamp module is completely disabled.

Created scripts to add underground biomes for snowy and sandy biometypes.

Adjust the Better Caves config so that cascading gravel is enabled again.