Very large update, read description
Haven't touched this pack in a really long time because I was busy. I was bored this week so I decided to get the Github version current so I could download it and play on my laptop lmao.

Added a few new mods, see commit

Updated dozens of mods

Changed dozens of configs

Changed/ added a few scripts

Added a plethora of new textures